requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.4
django-allauth==65.3.1
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.4
django-crispy-forms==1.14.0
pydantic==2.10.5
pydantic_core==2.27.2
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
